finished WalletWindowControler class.

+++ src/main/java/pl/lodz/uni/math/app/model/dao/WalletDAODatabaseImpl.java
@@ -4,10 +4,13 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import pl.lodz.uni.math.app.model.domain.Category;
 import pl.lodz.uni.math.app.model.domain.Wallet;
 
 public class WalletDAODatabaseImpl implements WalletDAO {
@@ -15,6 +18,8 @@ public class WalletDAODatabaseImpl implements WalletDAO {
 	private static final Logger log = LogManager.getLogger(WalletDAODatabaseImpl.class);
 
 	private static final String WALLET = "Wallet";
+
+	private static final String WALLET_ID = "wallet_id";
 
 	private Connection connection = null;
 
@@ -86,17 +91,19 @@ public class WalletDAODatabaseImpl implements WalletDAO {
 	@Override
 	public boolean removeWallet(Wallet wallet) {
 		if (wallet != null) {
-			String sql = "DELETE FROM " + WALLET + " WHERE id=?";
-			PreparedStatement statement = null;
-			try {
-				statement = connection.prepareStatement(sql);
-				statement.setInt(1, wallet.getId());
-				int result = statement.executeUpdate();
-				statement.close();
-				connection.commit();
-				return (result == 0) ? false : true;
-			} catch (SQLException e) {
-				System.out.println("Error ocurred in removeWallet method. Message: " + e.getMessage());
+			if (removeOperationFromOperationTableWhereWalletIs(wallet)) {
+				String sql = "DELETE FROM " + WALLET + " WHERE id=?";
+				PreparedStatement statement = null;
+				try {
+					statement = connection.prepareStatement(sql);
+					statement.setInt(1, wallet.getId());
+					int result = statement.executeUpdate();
+					statement.close();
+					connection.commit();
+					return (result == 0) ? false : true;
+				} catch (SQLException e) {
+					System.out.println("Error ocurred in removeWallet method. Message: " + e.getMessage());
+				}
 			}
 		}
 		return false;
@@ -134,6 +141,45 @@ public class WalletDAODatabaseImpl implements WalletDAO {
 			}
 		} catch (SQLException e) {
 			log.error("Error ocurred in checkIfWalletExistInDatabase method. Message: " + e.getMessage());
+		}
+		return false;
+	}
+
+	@Override
+	public List<Wallet> getWallets() {
+		List<Wallet> returnList = new ArrayList<>();
+		String sql = "SELECT * FROM " + WALLET;
+		PreparedStatement statement = null;
+		try {
+			statement = connection.prepareStatement(sql);
+			ResultSet resultSet = statement.executeQuery();
+			while (resultSet.next()) {
+				returnList.add(new Wallet(resultSet.getInt(1), resultSet.getString(2)));
+			}
+		} catch (SQLException e) {
+			log.error("Error ocurred in getWallets method. Message: " + e.getMessage());
+		}
+		return returnList;
+	}
+
+	/*
+	 * this method removes dependencies from Operation table.
+	 */
+	private boolean removeOperationFromOperationTableWhereWalletIs(Wallet wallet) {
+		if (wallet != null) {
+			String sql = "DELETE FROM " + WALLET + " WHERE " + WALLET_ID + "=?";
+			PreparedStatement statement = null;
+			try {
+				statement = connection.prepareStatement(sql);
+				statement.setInt(1, wallet.getId());
+				statement.executeUpdate();
+				statement.close();
+				connection.commit();
+				return true;
+			} catch (SQLException e) {
+				log.error("Error ocurred while removing operations from Operation table, base on wallet_id. Message: "
+						+ e.getMessage());
+			}
 		}
 		return false;
 	}
